Moved validate logic to form requests and changed url array_filter

+++ app/Models/FriendGame.php
@@ -30,7 +30,7 @@ class FriendGame extends Pivot
     protected function casts(): array
     {
         return [
-            "rating" => "integer",
+            "rating" => "float",
         ];
     }
 }
